refactor(shared): extract Runtime port interfaces to L0 shared/runtime-ports.ts
Move all Runtime port interfaces (RuntimeTime, RuntimeStorage, RuntimePaths,
RuntimeProcess, RuntimeIds, RuntimeEnv, etc.) from L1 execution/runtime.ts
to L0 shared/runtime-ports.ts. Update 6 L0 modules (shared/utils, file-tail,
session-entry, fs-lock, infra/plugin-registry, backend-info) to import from
the new location. L1 runtime.ts re-exports for existing consumers.

Fixes layer direction violation: L0 no longer imports from L1.
