+++ src/shared/file-tail.ts
@@ -1,5 +1,5 @@
 import { closeSync, openSync, readSync, statSync } from 'node:fs';
-import type { RuntimeStoragePort } from '../execution/runtime.js';
+import type { RuntimeStoragePort } from './runtime-ports.js';
 
 export type FileTailChunk = {
   lines: string[];
